+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fix main branch remote tracking
 * Require the env, run, and verify commands to be ran from the project root directory.
 * Rename bin/conda to bin/env
+* Add -e,--environment options to `cap env` command
+* `cap env` shows all variables created in the CAPTURE environment setup
 
 :warning: **WARNING** This is a breaking change!
 * Add slurm option to `cap run`